feat: commit staged changes and push branch
- Added new files for graph discovery and network pages
- Removed large legacy graph page and replaced with redirect to network page

This commit includes a large refactor of the graph workspace UI, splitting discovery and network views into separate pages and cleaning up the main graph page to redirect to the network view.

No lockfile changes detected.
